// src/app/Interfaces/trabajador.ts
export interface Trabajador {
  idTrabajador: number;
  nombre: string;
  apellido: string;
  sexo: string;
  fechaNacimiento: string;
  idTipoDocumento: number;
  tipo: string;
  numeroDocumento: string;
  direccion: string;
  telefonoFijo: string;
  telefonoCelular: string;
  email: string;
  estado: string;
}